Add Enumeration::Value
